requirements.txt
# Core dependencies
mcp>=0.1.0
motor>=3.3.2
python-dotenv>=1.0.0
typing-extensions>=4.9.0
pydantic>=2.6.0
pymongo>=4.6.0

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# Development dependencies
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
flake8-docstrings>=1.7.0
pre-commit>=3.6.0